Use Int64 for Category primary key
Seems there is a bug in Avram:
https://github.com/luckyframework/avram/issues/249

+++ src/models/category.cr
@@ -2,7 +2,7 @@ class Category < BaseModel
   skip_default_columns
 
   table do
-    primary_key id : Int16
+    primary_key id : Int64
     column name : String
     column hashtag : String
     column slug : String
